+++ example/step_connect/step_connect.cpp
@@ -1,12 +1,14 @@
 /*
  * libasesame3btサンプル
- * SesameのBluetoothアドレスを指定して接続、入力されたコマンドを実行
+ * 接続フェーズをステップバイステップで実行する
  */
 #include <Arduino.h>
 #include <Sesame.h>
 #include <SesameClient.h>
+#include <inttypes.h>
 #include <libsesame3bt/util.h>
 #include <cctype>
+#include <optional>
 // Sesame鍵情報設定用インクルードファイル
 // 数行下で SESAME_SECRET 等を直接定義する場合は別ファイルを用意する必要はない
 #if __has_include("mysesame-config.h")
@@ -50,10 +52,18 @@
 using libsesame3bt::history_tag_type_t;
 using libsesame3bt::Sesame;
 using libsesame3bt::SesameClient;
+using libsesame3bt::core::result_t;
 namespace util = libsesame3bt::core::util;
+
+#define DEBUG_AUTH_ERROR 0
+#if DEBUG_AUTH_ERROR
+#undef SESAME_SECRET
+#define SESAME_SECRET "00000000000000000000000000000000"
+#endif
 
 SesameClient client;
 SesameClient::Status last_status;
+SesameClient::state_t sesame_state;
 
 static const char*
 motor_status_str(Sesame::motor_status_t status) {
@@ -135,7 +145,8 @@ receive_history(SesameClient& client, const SesameClient::History& history) {
 	gmtime_r(&history.time, &tm);
 	// 過去に通知されたものと同じものが通知される場合がある。record_idで重複を検査可能
 	Serial.printf(
-	    "History(%ld) type=%u, %04d/%02d/%02d %02d:%02d:%02d, tag(%u)=%s, history_tag_type=%s, s_volt=%s, s_volt2=%s, pct=%s, "
+	    "History(%" PRId32
+	    ") type=%u, %04d/%02d/%02d %02d:%02d:%02d, tag(%u)=%s, history_tag_type=%s, s_volt=%s, s_volt2=%s, pct=%s, "
 	    "pct(opensensor)=%s, extra=%s\n",
 	    history.record_id, static_cast<uint8_t>(history.type), tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday, tm.tm_hour, tm.tm_min,
 	    tm.tm_sec, history.tag_len, history.tag,
@@ -160,6 +171,11 @@ receive_registered_devices(SesameClient& client, const std::vector<SesameClient:
 		Serial.printf("uuid=%s, os=%s\n", NimBLEUUID(dev.uuid, sizeof(dev.uuid)).reverseByteOrder().toString().c_str(),
 		              os_str(dev.os_ver));
 	}
+}
+
+void
+state_update(SesameClient& client, SesameClient::state_t state) {
+	sesame_state = state;
 }
 
 void
@@ -205,7 +221,7 @@ setup() {
 	}
 	Serial.println("Sesame keys set");
 	// SesameClient状態コールバックを設定
-	client.set_state_callback([](auto& client, auto state) { Serial.printf("sesame state changed to %s\n", state_str(state)); });
+	client.set_state_callback(state_update);
 	// Sesame状態コールバックを設定
 	client.set_status_callback(status_update);
 	// 履歴受信コールバックを設定
@@ -219,173 +235,71 @@ setup() {
 
 enum class app_state { init, wait_connected, wait_running, running, done };
 
-static uint32_t last_tried;
-static app_state state = app_state::init;
-static bool input_discarded;
-static int try_count = 0;
-
 constexpr const char* MENU_STR = R"(
+C) Connect
+A) Start Authenticate
+D) Disconnect
+
 L) Lock
-U) Unlock)"
-#if USE_UUID_TAG
-                                 R"(
-T) Lock(TAG_UUID)
-V) Unlock(TAG_UUID))"
-#endif
-                                 R"(
-C) Click(tag)
-D) Click()
-0～9) Click(N)
-R) Request status
-H) Request history
+U) Unlock
 
 X) Exit
-
 input>>)";
+
+SesameClient::state_t last_state = sesame_state;
+bool wait_input;
 
 void
 loop() {
-	switch (state) {
-		case app_state::init:
-			if (last_tried == 0 || millis() - last_tried > 500) {
-				if (try_count > 4) {
-					Serial.println("Giving up");
-					state = app_state::done;
-					return;
-				}
-				try_count++;
-				Serial.println("Connecting async...");
-				last_tried = millis();
-				if (!client.connect_async()) {
-					Serial.println("Failed to connect async");
-					state = app_state::init;
-					return;
-				}
-				Serial.println("async_connect retruned");
-				state = app_state::wait_connected;
+	if (last_state != sesame_state) {
+		Serial.printf("sesame state = %s\n", state_str(sesame_state));
+		last_state = sesame_state;
+	}
+	if (!wait_input) {
+		Serial.print(MENU_STR);
+		wait_input = true;
+	}
+	int c = Serial.read();
+	if (c <= 0) {
+		delay(100);
+		return;
+	}
+	Serial.printf("%c", c);
+	c = std::tolower(c);
+	switch (c) {
+		case 'c':
+			Serial.println("connecting");
+			if (!client.connect_async()) {
+				Serial.println("Failed to connect_async");
 			}
 			break;
-		case app_state::wait_connected:
-			if (client.get_state() == SesameClient::state_t::connected) {
-				Serial.println("Connected");
-				last_tried = millis();
-				if (client.start_authenticate()) {
-					Serial.println("Started authenticate");
-					state = app_state::wait_running;
-				} else {
-					Serial.println("Failed to start authenticate");
-					client.disconnect();
-					state = app_state::init;
-					return;
-				}
-			} else if (client.get_state() != SesameClient::state_t::connecting) {
-				Serial.println("Failed to connect");
-				client.disconnect();
-				state = app_state::init;
-				return;
+		case 'a':
+			Serial.println("authenticating");
+			if (!client.start_authenticate()) {
+				Serial.println("Failed to start_authenticate");
 			}
 			break;
-		case app_state::wait_running:
-			if (client.is_session_active()) {
-				// SesameClientの状態が Sesame::state_t::active になると設定を読み出し可能
-				// 設定は接続完了時にのみ読み出されるため、接続後に他のアプリで変更された値を知ることはできない
-				// SESAME Bot とそれ以外では読み出される設定値クラスが異なる
-				// LockSettingとBotSettingのvariant型が得られるので、型を指定(判定)して読み出す必要がある
-				if (const auto* setting = std::get_if<SesameClient::LockSetting>(&client.get_setting()); setting) {
-					// 自動ロック時間は SESAME 5(PRO) のみ取得可能。その他の機種では -1 (不明)が得られる。
-					// auto_lock_sec() が 0 であれば自動ロック無効
-					Serial.printf("Setting lock=%d,unlock=%d,auto_lock=%d\n", setting->lock_position(), setting->unlock_position(),
-					              setting->auto_lock_sec());
-				} else if (const auto* setting = std::get_if<SesameClient::BotSetting>(&client.get_setting()); setting) {
-					Serial.printf(
-					    "Setting "
-					    "pref_dir=%u,lock_sec=%u,unlock_sec=%u,click_lock_sec=%u,click_unlock_sec=%u,click_hold_sec=%u,button_mode=%u\n",
-					    setting->user_pref_dir(), setting->lock_sec(), setting->unlock_sec(), setting->click_lock_sec(),
-					    setting->click_unlock_sec(), setting->click_hold_sec(), setting->button_mode());
-				} else {
-					Serial.println("This model has no setting");
-				}
-				Serial.print(MENU_STR);
-				state = app_state::running;
-				break;
-			} else if (client.get_state() == SesameClient::state_t::idle) {
-				Serial.printf("Disconnected");
-				state = app_state::done;
-				break;
+		case 'd':
+			Serial.println("disconnecting");
+			if (!client.disconnect()) {
+				Serial.println("Failed to disconnect");
 			}
 			break;
-		case app_state::running:
-			// 認証が完了した
-			if (client.is_session_active()) {
-				if (!input_discarded) {  // discard VSCode Terminal initial echo
-					while (Serial.read() >= 0) {
-					}
-					input_discarded = true;
-				}
-				int c = Serial.read();
-				if (c <= 0) {
-					break;
-				}
-				Serial.printf("%c", c);
-				c = std::tolower(c);
-				switch (c) {
-					case '0':
-					case '1':
-					case '2':
-					case '3':
-					case '4':
-					case '5':
-					case '6':
-					case '7':
-					case '8':
-					case '9':
-						client.click(c - '0');
-						break;
-					case 'l':
-						client.lock("施錠テスト");
-						break;
-					case 'u':
-						client.unlock("開錠テスト");
-						break;
-#if USE_UUID_TAG
-					case 't':
-						client.lock(TAG_TYPE, NimBLEUUID(TAG_UUID));
-						break;
-					case 'v':
-						client.unlock(TAG_TYPE, NimBLEUUID(TAG_UUID));
-						break;
-#endif
-					case 'c':
-						client.click("クリックテスト");
-						break;
-					case 'd':
-						client.click();
-						break;
-					case 'r':
-						client.request_status();
-						Serial.println("Status requested");
-						break;
-					case 'h':
-						client.request_history();
-						Serial.println("History requested");
-						break;
-					case 'x':
-						client.disconnect();
-						state = app_state::done;
-						break;
-				}
-				if (state == app_state::running) {
-					Serial.print(MENU_STR);
-				}
-			} else {
-				Serial.println("Disconnected");
-				state = app_state::done;
-				break;
+		case 'l':
+			Serial.println("locking");
+			if (!client.lock("")) {
+				Serial.println("Failed to lock");
 			}
 			break;
-		case app_state::done:
-			delay(1000);
+		case 'u':
+			Serial.println("unlocking");
+			if (!client.unlock("")) {
+				Serial.println("Failed to unlock");
+			}
 			break;
+		default:
+			Serial.println();
+			Serial.print(MENU_STR);
 	}
 	delay(100);
 }